requirements.txt
djangorestframework
drf_yasg
